Complete airline company service requirements.

+++ src/main/java/com/kurt/springboot/airline/dao/AirlineCompanyRepository.java
@@ -3,6 +3,6 @@ package com.kurt.springboot.airline.dao;
 import com.kurt.springboot.airline.model.AirlineCompany;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AirlineCompanyRepository extends JpaRepository<AirlineCompany, Integer> {
+public interface AirlineCompanyRepository extends JpaRepository<AirlineCompany, Long> {
 
 }
